Renamed TrelloExec to TrelloService and moved it to the Service package along with other Service classes.
